refactor and setup eslint
