+++ index.html
@@ -36,7 +36,7 @@
   <div id="banner">
     Use F/B/L/R/U/D to rotate cube faces. Hold shift for counter-rotation.
   </div>
-  <script type="module" src="/rubics_cube.js"></script>
+  <script type="module" src="src/rubics_cube.js"></script>
 </body>
 
 </html>
