+++ src/three-js-helper.js
@@ -1,0 +1,37 @@
+import * as THREE from "three";
+import { OrbitControls } from "three/examples/jsm/controls/OrbitControls";
+
+export function createCamera() {
+  const camera = new THREE.PerspectiveCamera(
+    75,
+    window.innerWidth / window.innerHeight,
+    0.1,
+    1000,
+  );
+  camera.position.set(5, 5, 5);
+  camera.lookAt(0, 0, 0);
+  return camera;
+}
+
+export function createRenderer() {
+  const renderer = new THREE.WebGLRenderer({ antialias: true });
+  renderer.setSize(window.innerWidth, window.innerHeight);
+  return renderer;
+}
+
+export function createLights() {
+  const group = new THREE.Group();
+  group.add(new THREE.AmbientLight(0xffffff, 0.5));
+  const dirLight = new THREE.DirectionalLight(0xffffff, 0.8);
+  dirLight.position.set(1, 1, 1);
+  group.add(dirLight);
+  return group;
+}
+
+export function createControls(camera, renderer) {
+  // OrbitControls
+  const controls = new OrbitControls(camera, renderer.domElement);
+  controls.enableDamping = true;
+  controls.dampingFactor = 0.05;
+  return controls;
+}
